Issue 37 kokkos storage

+++ src/datastream.hpp
@@ -50,7 +50,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Array<T,length,CheckingPol
 template<
   typename T,
   template<int> class CheckingPolicy,
-  template<typename, int> class StoragePolicy
+  template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 1, CheckingPolicy, StoragePolicy>&);
 
@@ -62,7 +62,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 1, CheckingPolicy,
 template<
   typename T,
   template<int> class CheckingPolicy,
-  template<typename, int> class StoragePolicy
+  template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 2, CheckingPolicy, StoragePolicy>&);
 
@@ -75,7 +75,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 2, CheckingPolicy,
 template<
   typename T,
   template<int> class CheckingPolicy,
-  template<typename, int> class StoragePolicy
+  template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 3, CheckingPolicy, StoragePolicy>&);
 
@@ -88,7 +88,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 3, CheckingPolicy,
 template<
   typename T,
   template<int> class CheckingPolicy,
-  template<typename, int> class StoragePolicy
+  template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 4, CheckingPolicy, StoragePolicy>&);
 
@@ -102,7 +102,7 @@ std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 4, CheckingPolicy,
 template<
   typename T,
   template<int> class CheckingPolicy,
-  template<typename, int> class StoragePolicy
+  template<typename, size_t> class StoragePolicy
 >
 std::ostream &operator<<(std::ostream&, const schnek::Grid<T, 5, CheckingPolicy, StoragePolicy>&);
